针对Windows运行进行的修改

+++ external/landmark_detection/FaceBoxesV2/utils/nms_wrapper.py
@@ -7,9 +7,11 @@
 
 from .nms.cpu_nms import cpu_nms, cpu_soft_nms
 
+import numpy as np 
+
 def nms(dets, thresh):
     """Dispatch to either CPU or GPU NMS implementations."""
 
     if dets.shape[0] == 0:
         return []
-    return cpu_nms(dets, thresh)
+    return cpu_nms(dets.astype(np.float32), float(thresh))
